Move toml file settings struct out of `check.rs`
